Implemented a mechanism to send web requests to TRIPS and parse the reply into LogicalForm. Designed an XML template for verbal commands and wrote an example.
TODO: Write the command template parser. Probably skip all the extra validation, but leave markers to come back in the future. Implement a mechanism to compare Logical forms for a match.

TODO: After that, put together a pipeline of listen -> parse -> match. For now just have it output the name of matched command and bound parameters.